requirements.txt
google-auth==2.30.0
pandas==2.0.1
boto3==1.34.149
prettytable==3.10.2
pandas-gbq==0.22.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
pyarrow==17.0.0